requirements.txt
Flask==3.1.0
PyYAML==6.0.2
Requests==2.32.3
